requirements.txt
allure-pytest
pytest
requests
selenium
urllibs